Implement call streaming notification.
- Fixing some issues with CallStreamingController:
1. onCallStreamingStopped was not called when streaming stops.
2. added missing break for some cases statements.
- Fix bug where transactional calls don't set the provided display name
on the call.  This meant that the caller name wouldn't show up on BT, or
in this case the call streaming notification.
- Added new CallStreamingNotification class; this just listens for calls
that are streaming and posts or removes a call streaming notification.
- Added code in TelecomBroadcastIntentProcessor to handle hangup and stop
streaming from the call streaming notification.
- Fix bug in transactional test app which meant that all calls would be
treated as incoming.
- Add a call streaming test app which implements a bare minimum streaming
service for test purposes.

Test: Manual testing using test app.
Bug: 277232336
Merged-In: Id09ba876bc958e5f4f0f560186035f293ba4dfc5
Change-Id: Id09ba876bc958e5f4f0f560186035f293ba4dfc5