Prevent Remote Connection Services from unbinding when conference merges.
When a remote connection service conference call is initiated (such as a Fi conference call), two instances of ConnectionServiceWrapper are created (TelephonyConnectionService (which is the RCS) and the primary connectiong service (such as Fi)). When CSW#addConferenceCall and CSW#addExistingConnection invoked Call#replaceConnectionService for the RCS instance of CSW, it would occasionally decrement the associated call count below 0 and cause the ServiceBinder to unbind, dropping all calls.

This CL ensures invocations of CSW#addConferenceCall and CSW#addExistingConnection by a remote connection service are ignored and thus no errant unbinding can occur.

Fixes: 271631732
Test: manually merge a conference call via Tycho; ensure no calls drop
Change-Id: I3cab8afb4505a0a7826831e16d89a4216940a408