Generate an AnomalyReport when CsFcsMgr times out
A bugreport came in a user tried to answer an incoming call and the
notification froze. This was due to the user spamming the button and the
connection service focus manager timing out ever time the
getCurrentFocusCall was queried.

It'd be insightful to see how many times this is happening to determine
the importance in a refactor and better understand the issue.

Changes:
- remove the redundant log in CallsManager that fetches the active call
- dump the handler when the timeout occurs
- report the anomaly when the timeout occurs

Bug: 280128990
Test: manual (generate BR and inspect)
Change-Id: Ibe9130e795daa0bcceb34c7ac7afeaef88c96984