Add intent filter priority for critical intents Telecom relies on.
We saw an issue where the Bluetooth ACTION_ACTIVE_DEVICE_CHANGED intent
was not received by Telecom in a timely manner.  As a consequence Telecom
could not route audio to the BT device.

Adding in intent filter priority for all intents Telecom registers to
ensure that it can receive timely notification.

Test: Manual regression testing of calling with bluetooth.
Bug: 284410158
Change-Id: I46a259c6cb8c8c1a3f25460d6cd9075210c3d8a6