Resolve CallEndpoint nullability in BT audio route
There was an issue which came up where the endpoint name received in
HeadsetMediaButton#onCallEndpointChanged was null. CallEndpoints are
expected to hold non-null names. In the bug report, we see that the
error was invoked coming from the ActiveBluetoothRoute and looking at
the instantation of the endpoint for BT, we find two binder calls to
retrieve the bluetooth device name. It is likeley that the 2nd binder
call fails, which is resulting in the null value here. The resolution is
to remove the redundant binder call and instead store the value returned
from the first invocation.

Fixes: 286310499
Test: atest TelecomUnitTests
Change-Id: I2338d6675cb401f237a4022a5afcd9a5a4328846