isInSelfManagedCall for (package,user) should return true for calls being setup
Duo reported a regression that outgoing and incoming call notifications
became dismissiable. Upon investigation, this turned out to be a timing
issue where NotificationManagerService was querying Telecom to determine
if a self-managed call was added but Telecom had not yet added the call
to tracking.

The fix is to cover the time window where Telecom is first made aware of
the call in CallsManager until the call is offically added to
CallsManager. The time window is almost at a second which is a
significant amount of time.

Fixes: 282997779
Test: 4 new unit tests + manual testing:
      (m1) Flashing Duo's latest apk and verifying INCOMING call
           notifications are NON-dismissable
      (m2) Flashing Duo's latest apk and verifying OUTGOING call
           notifications are NON-dissmissable
      (m3) Flashing Duo's latest apk and verifying ONGOING call
           notifications are NON-dissmissable
      (m4) Testing the above cases on Telecom Transactional VoIP
	   app to ensure the change is general to all VoIP apps

Change-Id: If0cbbec59410824e0c00e51d4db3de571dcd41f2