Resolve MT Work Profile ECall
Currently, we don't properly support handling work calls across profile
boundaries. This only applies to emergency calling use cases as
non-emergency work calls are always rejected across profiles. The
existing check to verify if the work profile is paused only applies to
the context of an admin user using the work sim. However, it is possible
that the user is signed in as a guest/secondary user, the profile isn't
paused, and they receive an incoming emergency call on the work sim.

In order to account for this, we also need to ensure that if the current
user isn't the admin user, we should always perform the binding on that
user. In CallsManager, we need to ensure that the non-emergency work
calls are properly being rejected in the guest/secondary user. For the
ECBM use case, we need to ensure that all non-emergency calls are
permitted within the 5 minute time period.

Fixes: 287687156
Test: Added unit tests
Test: Manual verification on emergency calling use cases
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:01ee614f3b811015f125f6d083e8b779cdd241ff)
Merged-In: I924f0fb47952d9e68bc278e8f11fd5fe2541f289
Change-Id: I924f0fb47952d9e68bc278e8f11fd5fe2541f289