Fallback to QuiescentBluetoothRoute when abandon focus from ActiveBluetoothRoute.
Currently we try to reinitialized the call audio state machine
when we abandon the focus from ActiveBluetoothRoute. Since at that
time the BT audio already became inactive (while keep connected),
we mistakenly entered QuiescentEarpieceRoute which makes the audio
of the next call routed to earpiece route at the beginning. Fix
this by specific switch to QuiescentBluetoothRoute in this case.

Bug: 282069439
Test: atest CallAudioRouteStateMachineTest
Change-Id: I5a610b48fb0ca5f881df0b9319c3056d7dfe379f