specify what foregroundServiceTypes a client should be granted in VCM
Audio made some changes to enforce stricter mic access when an app is in
the background. This lead to all VoIP apps being suppressed since
Telecom was not specifying the foreground service type when granting
FGS delegation.

To adjust for the recent audio changes and foreground service changes,
Telecom needs to specify the foregroundServiceTypes on behalf of a
client when granting FGS delegation.  Otherwise, clients will experience
issues like audio or camera suppression when backgrounded.

Test: manual tests:
      (1) MIC is silenced when app w/ no mic perms is backgrounded
      (2) MIC is NOT silenced when app has mic perms is backgrounded
Bug: 288477797

Change-Id: Id6f85c24c59e045891e1c526cae99064800f4530