Fix crash during outgoing call + incoming SM call
When an outgoing call is in SELECT_PHONE_ACCOUNT state and an
incoming Self-Managed call is received, it will cause a crash
because we assume that the target PhoneAccount is set.

Instead, only count unholdable calls using Calls that have had
their target PhoneAccount set. This convention is set because
calls are not considered "active focus" until they have had their
PhoneAccount set. Otherwise, we could show UX to end the outgoing
call to accept the incoming  call, which will put Telecom in a
bad state because it does not properly end the call in
SELECT_PHONE_ACCOUNT state.

Bug: 285161098
Test: place outgoing managed call and put into SELECT_PHONE_ACCOUNT
and then create incoming Self-Managed call.
Test: atest TelecomUnitTests

Change-Id: Ica350292275a75936a3b529111ee0c3527d9dd8a