Fix ringtone play issue via LE devices.
Currently we query if the inband ringtone is enabled using audio
connected BT device. While when we try to decide if we should
connect to a BT device for ringtone, we haven't connect to any BT
device so the result of inband ring enabled in this case will
always be false. Change to use the most recently reported active
device to query if the inband ring is enabled or not since we'll
use the address of this device later to establish the audio
connection.

Bug: 278835031
Test: BluetoothDeviceManagerTest
Change-Id: I1188b75a347916b6dfda5cdacfcaf6812a5b96cd