Move set speakerphone out of async task executor.
This turned out to cause a regression related to switching to hearing
aids.  There are other set/clearcommunicationDevice operations elsewhere
in Telecom which are not part of the async task executor so we were seeing
race conditions where the clear hear doesn't happen before the set
in the Telecom BluetoothDeviceManager code.

For now we'll move this back out of the task executor since that was
only done to try and satisfy issues with Cuttlefish tests.

Test: Run CTS/Unit test suite.
Bug: 281934614
Change-Id: I9cc11eb11b59eb04b706a4bd6f61c1b7240a50c2