fix requestCallEndpointChange 2nd+ request to BT route
A bug was reported that requesting to switch to the BT route a 2nd time
would not trigger the onCurrentEndpointChanged callback.  Upon
inspection the request was not actually setting the aduio route and
short circuiting a success msg. This was ultimately due to the
isCurrentEndpointRequestedEndpoint method not enforcing a stricter
conditional check.

The fix is to ensure that the routes are the same before checking the bt
active device to deteremine if the user is requesting a new endpoint
that is already active. In short, the conditionals have been simplified.

Fixes: 290573705
Test: 2 new unit test + manual:
            (1) connect BT device to DUT
            (2) start MO using jetpack test app
            (3) switch CallEndpoint to EARPIECE
            (4) switch CallEndpoint back to BT
            expect: success; repeat steps 3+4 multi-times

Change-Id: I8d1c003f8269befc2cc26aafc60dfef3b80f70f3