Improve CallsLogManager logging code.
CallLogManager currently emits a Telecom event when we decide to log to
the call log.   To aid in diagnosis of potential problems, we track the
count of call log entries and max id before and after adding an entry to
help try to ascertain in Telecom whether logging did suceeed.  We will
also trigger an anomaly report if logging did not complete as expected to
help gain more information about this.

Also, added a button to the test app to add a load of call log entries to
help test cases where the call log has lots of stuff in it.

Test: Manual testing of new logging infra.
Bug: 290557274
Change-Id: Id7368331ba16682c1e61925bea7bf67859f135de