Add logging and anomaly detection in CallsLogManager.
CallLogManager currently emits a Telecom event when we decide to log to
the call log.   To aid in diagnosis of potential problems, we track the
count of call log entries and max id before and after adding an entry to
help try to ascertain in Telecom whether logging did suceeed.  We will
also trigger an anomaly report if logging did not complete as expected to
help gain more information about this.

Test: Manual testing for new logging infrastructure.
Bug: 294080756
Change-Id: I1b6d7e889f4bffd8119d027a02e51243b0f95c1c