add anom report when telecom requests the user but phony has default
Telephony allows users to select a default outgoing sim when placing
outgoing calls.  However, there has been cases where this was set and
Telecom still prompted the user. This is due to the fact that some
activities do not update Telecom and Telephony & Telecom get out of
sync. This new report should catch a case where Telecom prompts the user
to select a sim to place the call on but Telephony has a default

Fixes: 302397094
Test:  none
Change-Id: Iedd02f6c9b93e204e02f68e88fb56eff9083a30d