Restructure new outgoing call broadcast logic.
There was an error in the initial implementation which means that
redirection should have happened but it was not. This was not caught until
late when postsubmit staging builds ran the CTS tests.

I have added legacyProcessCall which is the original behavior with no
flags, and made processCall encapsulate the new flagged behavior.  This
does dupe the code, but it makes it much clearer to read and provides a
simple path for removing the old code when we remove the flag.

Test: Re-ran Telecom CTS tests with isNewOutgoingCallBroadcastUnblocking
flag off.
Test: Re-ran Telecom CTS tests with isNewOutgoingCallBroadcastUnblocking
flag on.
Test: Re-ran NewOutoingCallIntentBroadcasterTests to verify there was no
behavioral regression
Fixes: 300724616

Change-Id: I9323a4c1e4257e34b50d64d3e59b7ab75432c809