Ensure that getCallLogStats references the current user.
This CL updates the logic in getCallLogStats to ensure that the current user is referenced when checking that the call was properly logged to the call log. This was causing false anomaly reporting when a user was making calls from their work profile. This CL also ensures this anomlay report uses a newly created UUID so that the old UUID that was being used can be remotely disabled to prevent false positives from flooding in while still allowing genuine anomlies triggered after this logic fix merges to go through.

Flag: Test debug logging only
Test: manual using TestDPC
Fixes: 303858592
Change-Id: Idc634efacff0144b5e09ef6247fb8521200c2992