Allow return to call from keyguard UI.
When placing emergency calls from the keyguard UI and the device has a
lock, the user should be able to return to the call from the keyguard
UI at any time. Note that the keyguard UI belongs to the system user.

With the changes to handle multi-user calling scenarios,
a system user would not be able to access the available ICS for other
users (secondary/guest/work profiles), preventing the user from being
able to return to the call from the keyguard UI.

This CL ensures that the system user can access these calls with the
precondition that the device is currently locked and the ongoing call is
an emergency call. This also handles the work profile case when the
admin has also set a lock on the device and the user needs to be able to
return to the work profile emergency call.

Bug: 306582821
Test: atest TelecomUnitTests
Test: Manual to confirm that we are able to return to the ECC from the
keyguard UI for secondary/guest users as well as for work profiles.
Confirmed that the issue exists on the aforementioned profiles without
the change. For the work profile case, the user would need to place the
ECC when the device is unlocked and once it is locked, the user wouldn't
be able to return to the call.

Change-Id: I5d4b42fd162fa981c6f9cfe713815016ee0c5d92