Add flag for communication device refactor.
Add missing flag dependency for the call audio communication device
refactor work that was introduced as part of U. This CL also ensures
that the device set for communication is cleared when the audio mode is
set back to MODE_NORMAL (after call ends) and that the audio lost signal
is sent after the device has been cleared for communication. The latter
addresses issues with switching from a bluetooth LE audio device to
speaker to ensure that the device isn't considered an active device
still when switching audio routes.

Bug: 308968392
Test: atest TelecomUnitTests
Test: Manual to ensure that device set for communication is cleared
after a call ends.

Change-Id: I62a0d65c53b67a307b23d1015b994796e955a6aa