Clear locally tracked BT device for non-BT devices.
Ensures that the locally tracked BT device is cleared when a non-BT device is being set for communication. This occurred due to a race condition of multiple threads accessing setCommunicationDevice, which would be solved by using a lock. This just adds additional safety to ensure that BT device is indeed cleared when setting the communication device for non-BT devices.

Bug: 309922824
Test: atest TelecomUnitTests

Change-Id: I2f8e895e9e2d783935896c7b0a416b0917ff1e83