Ensure speaker on notification sent in legacy path
While flagging the communication device refactor work and in
reintroducing the legacy code path to set the speakerphone on/off, there
was a slight deviation of not properly updating the notification bar
status of the speakerphone when it is turned on.

Bug: 313922145
Test: Manual (confirmed that notification wasn't showing with
uncommitted changes and that issue is fixed with updated patch).

Change-Id: I7fb5fbd82a977f6c36f6028dca1199a00b7bf20d