Rename work profile associated users fix flag.
Rename the flag for the associated user fix that was addressed in U.
The original flagged change had to be reverted due to a build breakage
and was later relanded with the same flag name. Once the flag was
removed from Gerrit, the flag was deleted and is preventing the workflow
from progressing in Gantry. The fix is to create a separate flag in
order to begin a new workflow.

Bug: 315035693
Test: atest TelecomUnitTests
Change-Id: I573a9e6c7cf77e884c4c5c59f78af965e3c260da